// src/app/models/composicao-familiar.model.ts
export interface ComposicaoFamiliar {
    alfabetizado: boolean;
    estudaAtualmente: boolean;
    nivelSerieAtualConcluido: string;
    cursosTecnicoFormacaoProfissional: string;
    situacaoOcupacional: string;
    renda: string;
    aposentado: string;
    beneficio: string;
    deficiencia: string;
    problemaDeSaude: string;
    fazAlgumTratamento: boolean;
    fazAlgumTratamentoOnde: string;
    usaMedicamentoControlado: boolean;
    usaRecursosUbsLocal: boolean;
    trabalhoPastoralOuSocial: string;
    atividadeNaComunidadeSagradaFamilia: string;
    trabalhoVoluntario: string;
    trabalhoVoluntarioOnde: string;
}